added vanilla solid to examples
