--- examples/solid/src/entry-client.tsx
@@ -1,3 +1,0 @@
-import { StartClient, mount } from 'solid-start/entry-client'
-
-mount(() => <StartClient />, document)
