--- examples/solid/src/entry-server.tsx
@@ -1,9 +1,0 @@
-import {
-  StartServer,
-  createHandler,
-  renderAsync,
-} from 'solid-start/entry-server'
-
-export default createHandler(
-  renderAsync((event) => <StartServer event={event} />),
-)
