+++ examples/solid/src/index.tsx
@@ -1,0 +1,33 @@
+/* @refresh reload */
+import { render } from 'solid-js/web'
+
+import {
+  WagmiProvider,
+  configureChains,
+  createClient,
+  goerli,
+  mainnet,
+  publicProvider,
+} from 'wagmi-solid'
+
+import App from './App'
+
+const { provider, webSocketProvider } = configureChains(
+  [mainnet, goerli],
+  [publicProvider()],
+)
+
+const client = createClient({
+  autoConnect: true,
+  provider,
+  webSocketProvider,
+})
+
+render(
+  () => (
+    <WagmiProvider client={client}>
+      <App />{' '}
+    </WagmiProvider>
+  ),
+  document.getElementById('root') as HTMLElement,
+)
