--- examples/solid/src/routes/index.tsx
@@ -1,13 +1,0 @@
-// import { useConnect } from 'wagmi-solid'
-
-export default function Home() {
-  // const { connect } = useConnect()
-
-  return (
-    <main>
-      <p>aaaaa</p>
-      <p>nao carrega nada</p>
-      {/* <button onClick={() => connect()}>Connect</button> */}
-    </main>
-  )
-}
