+++ examples/solid/vite.config.ts
@@ -1,6 +1,12 @@
-import solid from 'solid-start/vite'
-import { defineConfig } from 'vite'
+import { defineConfig } from 'vite';
+import solidPlugin from 'vite-plugin-solid';
 
 export default defineConfig({
-  plugins: [solid()],
-})
+  plugins: [solidPlugin()],
+  server: {
+    port: 3000,
+  },
+  build: {
+    target: 'esnext',
+  },
+});
