+++ packages/solid/src/App.tsx
@@ -1,0 +1,8 @@
+function App() {
+  return (
+    <div>
+    </div>
+  );
+}
+
+export default App;
