+++ packages/solid/src/actions/package.json
@@ -1,0 +1,4 @@
+{
+  "type": "module",
+  "main": "../dist/actions.js"
+}
