+++ packages/solid/src/connectors/coinbaseWallet/package.json
@@ -1,0 +1,4 @@
+{
+  "type": "module",
+  "main": "../../dist/connectors/coinbaseWallet.js"
+}
