+++ packages/solid/src/connectors/injected/package.json
@@ -1,0 +1,4 @@
+{
+  "type": "module",
+  "main": "../../dist/connectors/injected.js"
+}
