+++ packages/solid/src/connectors/ledger/package.json
@@ -1,0 +1,4 @@
+{
+  "type": "module",
+  "main": "../../dist/connectors/ledger.js"
+}
