+++ packages/solid/src/connectors/metaMask/package.json
@@ -1,0 +1,4 @@
+{
+  "type": "module",
+  "main": "../../dist/connectors/metaMask.js"
+}
