+++ packages/solid/src/connectors/mock/package.json
@@ -1,0 +1,4 @@
+{
+  "type": "module",
+  "main": "../../dist/connectors/mock.js"
+}
