+++ packages/solid/src/connectors/package.json
@@ -1,0 +1,4 @@
+{
+  "type": "module",
+  "main": "../dist/connectors/index.js"
+}
