+++ packages/solid/src/connectors/walletConnect/package.json
@@ -1,0 +1,4 @@
+{
+  "type": "module",
+  "main": "../../dist/connectors/walletConnect.js"
+}
