+++ packages/solid/src/index.tsx
@@ -1,0 +1,7 @@
+/* @refresh reload */
+import { render } from 'solid-js/web';
+
+import './index.css';
+import App from './App';
+
+render(() => <App />, document.getElementById('root'));
