+++ packages/solid/src/providers/alchemy/package.json
@@ -1,0 +1,4 @@
+{
+  "type": "module",
+  "main": "../../dist/providers/alchemy.js"
+}
