+++ packages/solid/src/providers/infura/package.json
@@ -1,0 +1,4 @@
+{
+  "type": "module",
+  "main": "../../dist/providers/infura.js"
+}
