+++ packages/solid/src/providers/jsonRpc/package.json
@@ -1,0 +1,4 @@
+{
+  "type": "module",
+  "main": "../../dist/providers/jsonRpc.js"
+}
