+++ packages/solid/src/providers/public/package.json
@@ -1,0 +1,4 @@
+{
+  "type": "module",
+  "main": "../../dist/providers/public.js"
+}
